Refactors DuchessHistory and DuchessHistoryTest
