+++ src/main/java/Duke.java
@@ -1,6 +1,6 @@
 import duchess.exceptions.DuchessException;
 import duchess.logic.commands.Command;
-import duchess.model.task.DuchessHistory;
+import duchess.model.DuchessHistory;
 import duchess.parser.Parser;
 import duchess.storage.Storage;
 import duchess.storage.Store;
